backend/frontend/requirements.txt
flask
python-dotenv
requests
azure-cognitiveservices-vision-computervision
msrest